Extract SaveTool filename logic into dedicated module
Move build_save_filename, make_save_filename_hook, and
build_save_filename_from_cell out of plots.py and plot_grid_tabs.py
into a new save_filename.py module. Consolidate all related tests
into save_filename_test.py.

plots.py is already large and the filename logic is unrelated to
its core plotting responsibilities.

Prompt: "Should the helper functions really be on plots.py? The file
is already quite large and it seems they are quite unrelated to the
rest of the file? Same for tests."
